+++ Tests/Fluxor.UnitTests/SupportFiles/ThrowAggregateExceptionAction.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading;
+
+namespace Fluxor.UnitTests.SupportFiles
+{
+	public class ThrowAggregateExceptionAction
+	{
+		public ManualResetEvent TriggerHasFinished { get; } = new ManualResetEvent(false);
+	}
+}